refactor: centralize app management tooling
