+++ .github/workflows/apps-publish-submissions.yml
@@ -61,7 +61,7 @@ jobs:
           ISSUE_URL: ${{ github.event.issue.html_url }}
         run: |
           TODAY=$(date -u +%F)
-          APPROVED_DATE="$TODAY" node .github/scripts/app-validate-submission.js > "$RUNNER_TEMP/app-validation.json"
+          APPROVED_DATE="$TODAY" node apps/operation/app-management/validate-submission.js > "$RUNNER_TEMP/app-validation.json"
           jq -e '.valid == true' "$RUNNER_TEMP/app-validation.json" >/dev/null
           APP_JSON=$(jq -c .app "$RUNNER_TEMP/app-validation.json")
           {
@@ -87,8 +87,8 @@ jobs:
         run: |
           BRANCH="auto/app-${ISSUE_NUMBER}-${GITHUB_RUN_ID}-${GITHUB_RUN_ATTEMPT}"
           git switch -c "$BRANCH" origin/main
-          NEW_APP="$APP_JSON" node .github/scripts/app-prepend.js
-          node .github/scripts/app-update-greenhouse.js
+          NEW_APP="$APP_JSON" node apps/operation/app-management/prepend.js
+          node apps/operation/app-management/update-greenhouse.js
           git config user.name "$BOT_LOGIN"
           git config user.email "${BOT_ID}+${BOT_LOGIN}@users.noreply.github.com"
           git add apps/catalog.json apps/GREENHOUSE.md README.md
