+++ .github/workflows/apps-update-metrics.yml
@@ -39,10 +39,10 @@ jobs:
         env:
           GITHUB_TOKEN: ${{ steps.pollinations-ai.outputs.token }}
           TINYBIRD_READ_TOKEN: ${{ secrets.TINYBIRD_READ_TOKEN }}
-        run: node .github/scripts/app-update-metrics.js --verbose
+        run: node apps/operation/app-management/update-metrics.js --verbose
 
       - name: Regenerate greenhouse
-        run: node .github/scripts/app-update-greenhouse.js
+        run: node apps/operation/app-management/update-greenhouse.js
 
       - name: Check for changes
         id: changes
@@ -59,7 +59,7 @@ jobs:
           BRANCH="auto/app-metrics-$(date -u +%Y-%m-%d)"
           git checkout -b "$BRANCH"
           git add apps/catalog.json apps/GREENHOUSE.md README.md
-          git commit -m "chore: update app metrics, health, and greenhouse"
+          git commit -m "chore: update app metrics and greenhouse"
           git push -f origin "$BRANCH"
 
       - name: Create or update PR
@@ -76,7 +76,7 @@ jobs:
           else
             PR_URL=$(gh pr create \
               --title "chore: update app metrics $(date -u +%Y-%m-%d)" \
-              --body "Daily automated update of app metrics, health checks, and greenhouse")
+              --body "Daily automated update of app metrics and greenhouse")
             echo "Created PR: $PR_URL"
           fi
