+++ .github/workflows/ci-pull-request-checks.yml
@@ -25,7 +25,7 @@ jobs:
         run: npm ci
 
       - name: Validate app catalog
-        run: node .github/scripts/app-validate-catalog.js
+        run: node apps/operation/app-management/validate-catalog.js
 
       - name: Get changed files
         id: changed-files
@@ -62,11 +62,11 @@ jobs:
             echo "✅ All files are properly formatted"
           fi
 
-      - name: Get changed GitHub scripts
-        id: github-scripts
+      - name: Get changed automation scripts
+        id: automation-scripts
         run: |
-          FILES=$(git diff --diff-filter=d --name-only origin/${{ github.base_ref }}...HEAD | grep -E '^\.github/scripts/.*\.js$' || true)
-          echo "Changed GitHub scripts:"
+          FILES=$(git diff --diff-filter=d --name-only origin/${{ github.base_ref }}...HEAD | grep -E '^(\.github/scripts/|apps/operation/app-management/).*\.js$' || true)
+          echo "Changed automation scripts:"
           echo "$FILES"
           if [ -n "$FILES" ]; then
             echo "any_changed=true" >> $GITHUB_OUTPUT
@@ -74,14 +74,9 @@ jobs:
             echo "any_changed=false" >> $GITHUB_OUTPUT
           fi
 
-      - name: Test GitHub scripts
-        if: steps.github-scripts.outputs.any_changed == 'true'
-        run: |
-          if compgen -G ".github/scripts/*.test.js" > /dev/null; then
-            node --test .github/scripts/*.test.js
-          else
-            echo "No GitHub script tests found"
-          fi
+      - name: Test automation scripts
+        if: steps.automation-scripts.outputs.any_changed == 'true'
+        run: node --test .github/scripts/*.test.js apps/operation/app-management/*.test.js
 
       - name: Get changed TypeScript files in enter.pollinations.ai
         id: enter-ts-files
