+++ apps/operation/app-management/validate-catalog.js
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 
-const { CATALOG_FILE, readApps } = require("./lib/app-catalog.js");
+const { CATALOG_FILE, readApps } = require("./catalog.js");
 
 const apps = readApps();
 console.log(`Validated ${apps.length} apps in ${CATALOG_FILE}`);
